+++ Programmateur arrosage/include/myeeprom.h
@@ -1,20 +1,22 @@
 #ifndef MYEEPROM_h
 #define MYEEPROM_h
 
-const int mem_state = 17;
-const int mem_autostate = 12;
-const int mem_autoFreq = 11;
-const int mem_autoTimeOn = 10;
-const int mem_autoStartHour = 13;
-const int mem_autoStartMin = 14;
-const int mem_dayTemp = 150;
-const int mem_dayHumidity = 151;
-const int mem_autoSeason = 160;
-const int mem_tempSeason = 161;
-const int mem_sumerTimeon = 162;
-const int mem_sumerFreq = 163;
-const int mem_winterTimeon = 164;
-const int mem_winterFreq = 165;
+#include <Arduino.h>
+
+const byte mem_state = 17;
+const byte mem_autostate = 12;
+const byte mem_autoFreq = 11;
+const byte mem_autoTimeOn = 10;
+const byte mem_autoStartHour = 13;
+const byte mem_autoStartMin = 14;
+const byte mem_dayTemp = 150;
+const byte mem_dayHumidity = 151;
+const byte mem_autoSeason = 160;
+const byte mem_tempSeason = 161;
+const byte mem_sumerTimeon = 162;
+const byte mem_sumerFreq = 163;
+const byte mem_winterTimeon = 164;
+const byte mem_winterFreq = 165;
 /* memory address
 0 : 20:90 auto time on for ev
 1 : 21:91 auto frequency for ev
@@ -29,14 +31,14 @@ const int mem_winterFreq = 165;
     163 frequency summer
     164 duration winter
     165 frequency winter*/
-    
+
 class MYEEPROM
 {
 private:
 public:
     MYEEPROM();
     void init_memory();
-    void write(int memAdress, int value);
-    int Read(int memAdress);
+    void write(byte memAdress, byte value);
+    byte Read(byte memAdress);
 };
 #endif
